+++ src/pyvutils/validators/__init__.py
@@ -5,6 +5,10 @@ from .currency import (
     valid_numeric_currency_code
 )
 
+from .files import (
+    valid_filename
+)
+
 from .numerics import (
     valid_numeric
 )