+++ src/pyvutils/validators/files.py
@@ -1,0 +1,5 @@
+from pathvalidate import is_valid_filename
+
+
+def valid_filename(filename: str) -> bool:
+    return is_valid_filename(filename)